requirements.txt
Flask==2.1.2
pymongo==4.1.1

python-dotenv~=0.20.0